Added support for AVX512 instructions.
Also added functionality to break up the Reorder stage of the pipeline into multiple stages. This spread the CPU load over multiple threads. More stages means more processor overhead, so the number of stages should be kept as low as possible. However this is necessary for many core cpus where the clock speed of a single CPU is not high enough to perform the reorder operation in a single thread.