Found and fixed a major bug where xGPU was being passed buffer offsets that were incorrect by a factor of 2. This resulted in the output spectrum not being correct in an accumulation period.
Had to make a few major modifications to detect this. The first was modifying the f_engine_simulator.cpp script to be able to produce the same values per accumulation period. This made it very clear what the output from xGPU would be. The next modification was a minor modification to rx_real_imag.py to display these expected values correctly.

The issue was eventually traced to line 66 of GPUWrapper.cpp. inPacket_cast->getBufferOffset()*NUM_CHANNELS_PER_XENGINE*NUM_ANTENNAS*NUM_TIME_SAMPLES required an extra *2 at the end of the line. I thought this offset was in multiples of 2 polarisation complex samples. Instead it was in multiples of samples only. This *2 takes into account the polarisations.

This caused buffer overflow issues with allocating space to the GPU. This was fixed by reducing the value  DEFAULT_ACCUMULATIONS_THRESHOLD by a third, reducing the amount of memory required to be allocated

Accumulation produces the same results. There seems to be a bug where it eventually goes out of sync, but I am not certain if this is in the f_engine_simulator or the X-Engine. The next commit should be the solution to this problem